+++ src/stock_broker_tw/__init__.py
@@ -1,2 +1,2 @@
 """stock_broker_tw_server package."""
-__version__ = "0.1.0"
+__version__ = "0.1.2"